updated for latest iOS 6 SDK

+++ UIKit/UIImage+SA_Additions.m
@@ -37,7 +37,7 @@
 }
 
 - (NSString *) description {
-	return [NSString stringWithFormat: @"<%@: 0x%x, (%.0f, %.0f)>", [self class], self, self.size.width, self.size.height];
+	return [NSString stringWithFormat: @"<%@: 0x%x, (%.0f, %.0f)>", [self class], (int) self, self.size.width, self.size.height];
 }
 
 - (void) drawInRect: (CGRect) rect withContentMode: (UIViewContentMode) mode {
